server/requirements.txt
flask
flask_cors
python-dotenv
requests
PyMuPDF
Levenshtein
google-generativeai
XlsxWriter
